# Optimize Docker Build Layers and Add Sudo Privileges for Fast-LLM Container