--- a/Dockerfile
+++ b/Dockerfile
@@ -19,13 +19,13 @@ WORKDIR /app
 ENV PYTHONPATH=/app:/app/Megatron-LM \
     PATH=$PATH:/home/fast_llm/.local/bin/
 
-# Compile the C++ extensions (fast_llm/csrc)
-COPY --chown=fast_llm ./fast_llm/csrc/ fast_llm/csrc/
-RUN make -C ./fast_llm/csrc/
-
 # Copy the dependency files and install dependencies
 COPY --chown=fast_llm setup.py setup.cfg pyproject.toml ./
 RUN PIP_NO_INPUT=1 pip3 install --no-cache-dir -e ".[CORE,OPTIONAL,DEV]"
+
+# Compile the C++ extensions (fast_llm/csrc)
+COPY --chown=fast_llm ./fast_llm/csrc/ fast_llm/csrc/
+RUN make -C ./fast_llm/csrc/
 
 # Copy the rest of the code
 COPY --chown=fast_llm ./Megatron-LM Megatron-LM